+++ tests/Unit/RestApi/Device/DeviceManagement.spec.ts
@@ -22,6 +22,7 @@ describe('DeviceManagement', () => {
 		'model': 'rk30sdk-PN_B_series',
 		'serialNumber': '8F047121',
 		'brand': 'Sony',
+		'osVersion': '1.2.3',
 		'timezone': null,
 		'organizationUid': 'f4dc88XXXXXXfba4c',
 		'networkInterfaces': {